# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Login user",
        "tags": [
          "users"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create new user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a user by ID",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a user",
        "tags": [
          "users"
        ]
      }
    },
    "/properties": {
      "get": {
        "operationId": "PropertiesController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all properties",
        "tags": [
          "properties"
        ]
      },
      "post": {
        "operationId": "PropertiesController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create new property",
        "tags": [
          "properties"
        ]
      }
    },
    "/properties/{id}": {
      "get": {
        "operationId": "PropertiesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get property by id",
        "tags": [
          "properties"
        ]
      },
      "put": {
        "operationId": "PropertiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update property",
        "tags": [
          "properties"
        ]
      },
      "delete": {
        "operationId": "PropertiesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete property",
        "tags": [
          "properties"
        ]
      }
    },
...
    "/payments": {
      "get": {
        "operationId": "PaymentsController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all payments",
        "tags": [
          "payments"
        ]
      },
      "post": {
        "operationId": "PaymentsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Submit new payment",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get payment by ID",
        "tags": [
          "payments"
        ]
      },
      "put": {
        "operationId": "PaymentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update payment status",
        "tags": [
          "payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "description": "The role of the requesting user (admin, warden, owner, tenant)",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "The ID of the requesting user",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete payment record",
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "PG Rental Management API",
    "description": "The PG Rental Management System using in-memory structures",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "role",
          "username",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "property": {
            "type": "string"
          }
        }
      },
      "CreatePropertyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "rentMin": {
            "type": "number"
          },
          "rentMax": {
            "type": "number"
          },
          "rooms": {
            "type": "string"
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "location",
          "owner",
          "rentMin",
          "rentMax",
          "rooms",
          "amenities"
        ]
      },
      "UpdatePropertyDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "docsVerified": {
            "type": "boolean"
          },
          "inspectionPassed": {
            "type": "boolean"
          }
        }
      },
...
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "tenant": {
            "type": "string"
          },
          "property": {
            "type": "string"
          },
          "room": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "method": {
            "type": "string"
          },
          "transactionId": {
            "type": "string"
          }
        },
        "required": [
          "tenant",
          "property",
          "room",
          "amount",
          "method",
          "transactionId"
        ]
      },
      "UpdatePaymentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "clearance": {
            "type": "string"
          }
        }
      },
...
    }
  }
}